+++ wallet/core/src/account/variants/bip32watch.rs
@@ -162,8 +162,8 @@ impl Account for Bip32Watch {
         self
     }
 
-    fn sig_op_count(&self) -> u16 {
-        u16::try_from(self.xpub_keys.len()).unwrap()
+    fn sig_op_count(&self) -> u8 {
+        u8::try_from(self.xpub_keys.len()).unwrap()
     }
 
     fn minimum_signatures(&self) -> u16 {
