+++ wallet/core/src/account/variants/multisig.rs
@@ -191,8 +191,8 @@ impl Account for MultiSig {
         self
     }
 
-    fn sig_op_count(&self) -> u16 {
-        u16::try_from(self.xpub_keys.len()).unwrap()
+    fn sig_op_count(&self) -> u8 {
+        u8::try_from(self.xpub_keys.len()).unwrap()
     }
 
     fn minimum_signatures(&self) -> u16 {
